Adding files to enable library editing
from google:

Here's how to use the Arduino IDE to edit a library:
Add a dummy .ino file to the folder containing the library files you want to edit with the Arduino IDE. The .ino filename must match the folder name. This allows the Arduino IDE to open the library source files as if they were a sketch. This file will not actually be compiled as part of the library. I use it as a place to store my "to do" list. You could also use it for a test sketch for the library, in this way you can actually compile the library right from the same IDE window where you are editing it instead of having a sketch open in a separate IDE window.
Add a file named .development to the root of the library folder. This is necessary because otherwise the Arduino IDE treats the contents of the library as read-only (to prevent people from accidentally modifying example sketches).
If it doesn't have one already, add a file named library.properties in the root of the library folder. You can find the specification of the format here: https://arduino.github.io/arduino-cli/library-specification/. This is required for the .development feature to work. The exact contents of the library.properties file is not important, as long as it meets the specification. You could even just copy one from another library if you like, though I would recommend changing the name value so to avoid confusion.